requirements.txt
click==8.0.4
frozendict==2.3.7
more-itertools==9.1.0
numpy==1.23.5
pandas==2.0.0
plotly==5.14.1
pyyaml==6.0
ray[rllib]==2.1.0
tensorflow==2.12.0